# docs/openapi.yaml
openapi: 3.0.3
info:
  title: zk-id Demo API
  version: 0.6.0
  description: |
    Demo REST API for zk-id verification flows.

    This API implements the zk-id protocol version **zk-id/1.0-draft**.
    Protocol version is communicated via the `X-ZkId-Protocol-Version` header.
    Servers may reject incompatible protocol versions with `400`.

    This is a reference schema and may evolve as the protocol matures.
servers:
  - url: http://localhost:3000
paths:
...
  /api/challenge:
    get:
      summary: Get server-issued nonce and timestamp
      parameters:
        - $ref: '#/components/parameters/ProtocolVersionHeader'
      responses:
        '200':
          description: Challenge payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Challenge'
          headers:
            X-ZkId-Protocol-Version:
              $ref: '#/components/headers/ProtocolVersion'
        '400':
          description: Incompatible protocol version
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/issue-credential:
    post:
      summary: Issue a signed credential (Ed25519)
      parameters:
        - $ref: '#/components/parameters/ProtocolVersionHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [birthYear, nationality]
              properties:
                birthYear:
                  type: integer
                nationality:
                  type: integer
                userId:
                  type: string
      responses:
        '200':
          description: Issued credential
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  credential:
                    $ref: '#/components/schemas/SignedCredential'
          headers:
            X-ZkId-Protocol-Version:
              $ref: '#/components/headers/ProtocolVersion'
        '400':
          description: Bad request (including incompatible protocol version)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/verify-age:
    post:
      summary: Verify an age proof
      description: |
        Accepts a ProofResponse with claimType `age` or `age-revocable`.
        When claimType is `age-revocable`, the server also validates the Merkle root
        against the valid-credential tree and optional staleness policy.
        See `docs/schemas/proof-response.json` for the JSON schema.
      parameters:
        - $ref: '#/components/parameters/ProtocolVersionHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProofResponse'
      responses:
        '200':
          description: Verification result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerificationResult'
          headers:
            X-ZkId-Protocol-Version:
              $ref: '#/components/headers/ProtocolVersion'
        '400':
          description: Bad request (including incompatible protocol version)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/verify-nationality:
    post:
      summary: Verify a nationality proof
      description: |
        Accepts a ProofResponse with claimType `nationality`.
        See `docs/schemas/proof-response.json` for the JSON schema.
      parameters:
        - $ref: '#/components/parameters/ProtocolVersionHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProofResponse'
      responses:
        '200':
          description: Verification result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerificationResult'
          headers:
            X-ZkId-Protocol-Version:
              $ref: '#/components/headers/ProtocolVersion'
        '400':
          description: Bad request (including incompatible protocol version)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/verify-voting-eligibility:
    post:
      summary: Verify voting eligibility scenario (age + nationality)
      description: |
        Verifies the VOTING_ELIGIBILITY_US scenario: age >= 18 AND nationality = USA.
        Accepts an array of ProofResponse objects (one for age, one for nationality).
      parameters:
        - $ref: '#/components/parameters/ProtocolVersionHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScenarioProofRequest'
      responses:
        '200':
          description: Scenario verification result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScenarioVerificationResult'
          headers:
            X-ZkId-Protocol-Version:
              $ref: '#/components/headers/ProtocolVersion'
        '400':
          description: Bad request (including incompatible protocol version)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/verify-senior-discount:
    post:
      summary: Verify senior discount scenario (age >= 65)
      description: |
        Verifies the SENIOR_DISCOUNT scenario: age >= 65.
        Accepts an array of ProofResponse objects (single proof for age).
      parameters:
        - $ref: '#/components/parameters/ProtocolVersionHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScenarioProofRequest'
      responses:
        '200':
          description: Scenario verification result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScenarioVerificationResult'
          headers:
            X-ZkId-Protocol-Version:
              $ref: '#/components/headers/ProtocolVersion'
        '400':
          description: Bad request (including incompatible protocol version)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/verify-scenario:
    post:
      summary: Verify a scenario bundle (multiple proofs)
      description: |
        Verifies a scenario bundle containing multiple proofs for combined claims.
        Scenarios are pre-defined real-world use cases like voting eligibility or senior discounts.
      parameters:
        - $ref: '#/components/parameters/ProtocolVersionHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [scenarioId, response]
              properties:
                scenarioId:
                  type: string
                  description: The scenario identifier (e.g., "voting-eligibility-us")
                response:
                  type: object
                  description: Scenario-specific response containing proofs and metadata
      responses:
        '200':
          description: Verification result
          content:
            application/json:
              schema:
                type: object
                properties:
                  verified:
                    type: boolean
                  message:
                    type: string
          headers:
            X-ZkId-Protocol-Version:
              $ref: '#/components/headers/ProtocolVersion'
        '400':
          description: Bad request (including incompatible protocol version)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/revoke-credential:
    post:
      summary: Revoke a credential by ID
      parameters:
        - $ref: '#/components/parameters/ProtocolVersionHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [credentialId]
              properties:
                credentialId:
                  type: string
      responses:
        '200':
          description: Revocation result
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
          headers:
            X-ZkId-Protocol-Version:
              $ref: '#/components/headers/ProtocolVersion'
        '400':
          description: Bad request (including incompatible protocol version)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Credential not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/revocation/root:
    get:
      summary: Get current revocation root info with versioning and TTL
      description: |
        Returns the current valid-set Merkle root along with version metadata
        and caching guidance. Clients should cache the response for `ttlSeconds`
        and re-fetch before `expiresAt`. See PROTOCOL.md §Revocation Root Distribution.
      parameters:
        - $ref: '#/components/parameters/ProtocolVersionHeader'
      responses:
        '200':
          description: Current revocation root info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RevocationRootInfo'
          headers:
            X-ZkId-Protocol-Version:
              $ref: '#/components/headers/ProtocolVersion'
            Cache-Control:
              description: Cache guidance based on configured TTL (e.g., "public, max-age=300")
              schema:
                type: string
        '400':
          description: Bad request (including incompatible protocol version)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  parameters:
    ProtocolVersionHeader:
      name: X-ZkId-Protocol-Version
      in: header
      required: false
      description: Protocol version for compatibility (e.g., "zk-id/1.0-draft")
      schema:
        type: string
  headers:
    ProtocolVersion:
      description: Protocol version (e.g., "zk-id/1.0-draft")
      schema:
        type: string
  schemas:
...
    Challenge:
      type: object
      required: [nonce, requestTimestamp]
      properties:
        nonce:
          type: string
        requestTimestamp:
          type: string
          format: date-time
    RevocationRootInfo:
      type: object
      required: [root, version, updatedAt]
      properties:
        root:
          type: string
          description: Current Merkle root (decimal string)
        version:
          type: integer
          description: Monotonic root version (increments on each tree mutation)
        updatedAt:
          type: string
          format: date-time
          description: ISO 8601 timestamp of last tree mutation
        expiresAt:
          type: string
          format: date-time
          description: ISO 8601 timestamp after which this root should be re-fetched
        ttlSeconds:
          type: integer
          description: Recommended cache lifetime in seconds (default 300)
        source:
          type: string
          description: Identifier for the root source (issuer name, registry URL)
    Credential:
      type: object
      properties:
        id:
          type: string
        birthYear:
          type: integer
        nationality:
          type: integer
        salt:
          type: string
        commitment:
          type: string
        createdAt:
          type: string
          format: date-time
    SignedCredential:
      type: object
      properties:
        credential:
          $ref: '#/components/schemas/Credential'
        issuer:
          type: string
        signature:
          type: string
        issuedAt:
          type: string
          format: date-time
...
    ProofResponse:
      type: object
      required: [claimType, proof, nonce, requestTimestamp, signedCredential]
      properties:
        credentialId:
          type: string
        claimType:
          type: string
          enum: [age, nationality, age-revocable]
        proof:
          type: object
          properties:
            proof:
              $ref: '#/components/schemas/Proof'
            publicSignals:
              oneOf:
                - $ref: '#/components/schemas/AgePublicSignals'
                - $ref: '#/components/schemas/NationalityPublicSignals'
                - $ref: '#/components/schemas/AgeProofSignedPublicSignals'
                - $ref: '#/components/schemas/NationalityProofSignedPublicSignals'
                - $ref: '#/components/schemas/AgeProofRevocablePublicSignals'
        signedCredential:
          oneOf:
            - $ref: '#/components/schemas/SignedCredential'
            - $ref: '#/components/schemas/CircuitSignedCredential'
        nonce:
          type: string
        requestTimestamp:
          type: string
          format: date-time
...
    VerificationResult:
      type: object
      required: [verified]
      properties:
        verified:
          type: boolean
        claimType:
          type: string
        minAge:
          type: integer
        targetNationality:
          type: integer
        error:
          type: string
        protocolVersion:
          type: string
          description: Protocol version (e.g., "zk-id/1.0-draft")
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Error message
        details:
          type: string
          description: Additional error details
        verified:
          type: boolean
          description: Present in verification errors (always false)
    ScenarioProofRequest:
      type: object
      required: [proofs]
      properties:
        proofs:
          type: array
          items:
            $ref: '#/components/schemas/ProofResponse'
          description: Array of proof responses, one for each claim in the scenario
    ScenarioVerificationResult:
      type: object
      required: [verified, scenario, message]
      properties:
        verified:
          type: boolean
          description: Whether all scenario claims were verified successfully
        scenario:
          type: string
          description: Scenario name (e.g., "US Voting Eligibility")
        message:
          type: string
          description: Human-readable result message
        error:
          type: string
          description: Error message if verification failed
...